睡眠状态修改

+++ components/TIMER/include/TIMER.h
@@ -9,7 +9,7 @@ extern bool flag_collect;
 extern bool flag_cooldown;
 extern int cpm_bodyMove;
 extern int count_rate_bcg;
-extern float cpm_rate_bcg;
+extern double cpm_rate_bcg;
 extern int status;
 
 //ADC定时器，10ms，周期
